+++ addons/auth_totp_mail/data/mail_template_data.xml
